# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Crop:
    properties:
      category:
        type: string
      code:
        type: string
      created_at:
        type: string
      harvest_resource_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      yield_unit:
        type: string
    type: object
...
  domain.FieldCrop:
    properties:
      created_at:
        type: string
      crop_id:
        type: integer
      crop_name:
        type: string
      expected_yield_per_ha:
        type: number
      field_area_ha:
        type: number
      field_id:
        type: string
      field_name:
        type: string
      harvest_recorded_at:
        type: string
      harvest_recorded_quantity:
        type: number
      harvested_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      planted_area_ha:
        type: number
      planted_at:
        type: string
      production_total:
        type: number
      season_end:
        type: string
      season_id:
        type: integer
      season_name:
        type: string
      season_start:
        type: string
      updated_at:
        type: string
      yield_per_ha:
        type: number
      yield_unit:
        type: string
    type: object
...
  domain.Season:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      notes:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  domain.StockMovement:
    properties:
      actor_id:
        type: integer
      actor_name:
        type: string
      category:
        type: string
      created_at:
        type: string
      field_crop_id:
        type: integer
      field_operation_id:
        type: integer
      field_operation_label:
        type: string
      id:
        type: integer
      movement_type:
        $ref: '#/definitions/domain.StockMovementType'
      notes:
        type: string
      quantity_delta:
        type: number
      resource_id:
        type: integer
      resource_name:
        type: string
      resulting_quantity:
        type: number
      total_cost:
        type: number
      unit:
        type: string
      unit_cost:
        type: number
    type: object
  domain.StockMovementType:
    enum:
    - in
    - out
    - adjustment
    type: string
    x-enum-varnames:
    - StockMovementIn
    - StockMovementOut
    - StockMovementAdjustment
...
  handlers.cropRequest:
    properties:
      category:
        type: string
      code:
        type: string
      name:
        type: string
      notes:
        type: string
      yield_unit:
        type: string
    required:
    - name
    type: object
  handlers.fieldCropRequest:
    properties:
      crop_id:
        type: integer
      expected_yield_per_ha:
        type: number
      field_id:
        type: string
      harvested_at:
        type: string
      notes:
        type: string
      planted_area_ha:
        type: number
      planted_at:
        type: string
      production_total:
        type: number
      season_id:
        type: integer
    required:
    - crop_id
    - field_id
    - season_id
    type: object
  handlers.harvestResponse:
    properties:
      field_crop:
        $ref: '#/definitions/domain.FieldCrop'
      movement:
        $ref: '#/definitions/domain.StockMovement'
    type: object
...
  handlers.seasonRequest:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      notes:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - name
    - start_date
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST pentru managementul mașinilor agricole, operatorilor și asignărilor.
  title: Agri API
  version: "1.0"
paths:
...
  /crops:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Crop'
            type: array
      security:
      - BearerAuth: []
      summary: Culturi
      tags:
      - crops
    post:
      consumes:
      - application/json
      parameters:
      - description: Cultura
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.cropRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Crop'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Creare cultură
      tags:
      - crops
  /crops/{id}:
    delete:
      parameters:
      - description: ID cultură
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ștergere cultură
      tags:
      - crops
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID cultură
        in: path
        name: id
        required: true
        type: integer
      - description: Cultura
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.cropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Crop'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizare cultură
      tags:
      - crops
...
  /field-crops:
    get:
      parameters:
      - description: Filtru sezon
        in: query
        name: season_id
        type: integer
      - description: Filtru cultură
        in: query
        name: crop_id
        type: integer
      - description: Filtru teren
        in: query
        name: field_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.FieldCrop'
            type: array
      security:
      - BearerAuth: []
      summary: Culturi pe terenuri
      tags:
      - crops
    post:
      consumes:
      - application/json
      parameters:
      - description: Cultura pe teren
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.fieldCropRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.FieldCrop'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Creare cultură pe teren
      tags:
      - crops
  /field-crops/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ștergere cultură pe teren
      tags:
      - crops
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cultura pe teren
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.fieldCropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.FieldCrop'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizare cultură pe teren
      tags:
      - crops
  /field-crops/{id}/harvest:
    post:
      parameters:
      - description: ID cultură pe teren
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.harvestResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Recoltă în stoc
      tags:
      - crops
...
  /seasons:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Season'
            type: array
      security:
      - BearerAuth: []
      summary: Sezoane
      tags:
      - crops
    post:
      consumes:
      - application/json
      parameters:
      - description: Sezonul
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.seasonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Season'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Creare sezon
      tags:
      - crops
  /seasons/{id}:
    delete:
      parameters:
      - description: ID sezon
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ștergere sezon
      tags:
      - crops
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID sezon
        in: path
        name: id
        required: true
        type: integer
      - description: Sezonul
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.seasonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Season'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizare sezon
      tags:
      - crops
...
securityDefinitions:
  BearerAuth:
    description: 'Introdu token-ul cu prefixul "Bearer ": **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"